requirements.txt
Django==5.2
django-allauth==0.58.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.1
django-tinymce==4.0.0
django-filter==23.5
Pillow==10.3.0